Remove RegExpr from unit test
The unit test that verifies that less.render() throws an exception
when given invalid Less input no longer passes a RegExp argument
to Tape's throws() method.  Unfortunately, Tape's throws() method
only matches RegExps to thrown strings, but less.render() is now
throwing an object with a "message" property, and no RegExp will
match it.  So the best we can do here is not to try to match the
content of the exception at all.